Move markdown component into components subdir
